+++ cli/factory.rs
@@ -55,12 +55,12 @@ use std::collections::BTreeSet;
 use std::path::PathBuf;
 
 use deno_config::deno_json::ConfigFile;
-use deno_config::package_json::PackageJsonDepValue;
 use deno_config::workspace::PackageJsonDepResolution;
 use deno_config::workspace::WorkspaceResolver;
 use deno_core::error::AnyError;
 use deno_core::futures::FutureExt;
 use deno_core::FeatureChecker;
+use deno_package_json::PackageJsonDepValue;
 
 use deno_lockfile::WorkspaceMemberConfig;
 use deno_runtime::deno_fs;
