+++ ext/node/errors.rs
@@ -315,7 +315,7 @@ impl NodeJsErrorCoded for PathToDeclarationUrlError {
 pub struct PackageJsonLoadError(
   #[source]
   #[from]
-  pub deno_config::package_json::PackageJsonLoadError,
+  pub deno_package_json::PackageJsonLoadError,
 );
 
 impl NodeJsErrorCoded for PackageJsonLoadError {
