+++ ext/node/lib.rs
@@ -32,7 +32,7 @@ mod path;
 mod polyfill;
 mod resolution;
 
-pub use deno_config::package_json::PackageJson;
+pub use deno_package_json::PackageJson;
 pub use ops::ipc::ChildPipeFd;
 pub use ops::ipc::IpcJsonStreamResource;
 use ops::vm;
